libview: provide complete PpsSearchResult on search-context "result-activated"
This allows us to get direct access to the rectangles and directly go
to that result, instead of having to loop over every possible result.
In the process, it also becomes apparent that there's no need for
jump_to_find_page, since we know that the page we are getting does
have a result.

This is possible thanks to the previous commit, that limits the
location where these functions are called.