libdocument: add a way to get just a list of annotations
Instead of having to get a mapping. The mapping is an unnecessary
wrapper around the annotations consequence of faulty coordinate
systems being provided by default in the rectangle of
PopplerAnnot (see
https://gitlab.freedesktop.org/poppler/poppler/-/merge_requests/1652).

However, our PpsAnnotation interface as its own rectangle area, where
we save the rectangle from the mapping. Therefore, we can in theory
just use the annotation area as we currently use the mapping. That is
a follow-up cleanup, but for now add a new interface that just gets a
list of annotations, and that we can use to start replacing other
use-cases.