libview: do not transfer the opacity of the annotation to the window
This is a violation of the specification. Per the spec:

The constant opacity value that shall be used in
painting the annotation (see Sections 11.2, “Overview of Transparency,”
and 11.3.7, “Shape and Opacity Computations”). This value shall apply to
all visible elements of the annotation in its closed state (including its
background and border) but not to the pop-up window that appears when
the annotation is opened

Fixes #306