shell: add return values to signal callbacks
The signals in question do have a boolean return value
meant for handler propagation. Return true signals a
handled event, and false an unhandled one.

As this was not properly done in the C counterpart,
it would have been probably okay to return any value,
but we do need to, as the implicit void to false casting
C did (sadly) do for us luckily doesn't fly in Rust world.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>