shell: filter empty document titles
Some documents do not have a title, some PDF generators
instead of leaving out the field, do fill it with an
empty string instead. Here, filtering empty document
titles results in the same behaviour as with documents
without a title: Its file name gets used instead.

This used to be the case with the C version of the code.

Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/315

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>