shell: Make night mode a global flag
The night mode is not identical to inverted colors. It's a combination
of inverted colors and forced color schemes of libadwaita. We add a new
gsetting named dark-mode and use it to sync dark mode between
application instances. And the handle of this mode is also moved to
PpsWindow.

Fixes #303
Fixes #312

Signed-off-by: Qiu Wenbo <qiuwenbo@kylinos.com.cn>