libview: add pps_view_get_doc_point_for_page private API
This is intended to be used instead of
pps_view_get_mark_for_view_point, but for the cases where we actually
know the page. This has a bunch of internal uses, and could become
some useful API to extract state from a page in the future.

This also leads to quite some simplification in consumer APIs.

To make sure we keep the interpretation of the x, y values consistent
and meaning "a value in the widget coordinates", we need to move a
bunch of {x,y} += priv->scroll_{x,y} to inside get_doc_point_for_page
and find_page_at_location functions.