libview: remove view_point_to_doc_point
To reduce the complexity around it, we embed it in the last places
where it is still being used. The new transformation function
get_doc_point_for_page get it completely inlined, while the uses in
the selection discard the work related to rotations. This is due to
that code being already not functional when the document is
rotated. Any future work aimed at making that possible will anyway
require a sizable refactoring of all the selection code.