shell: document-view: fix condition for when to store show-sidebar setting
The preference to show the sidebar, aptly named
show-sidebar, is supposed to reflect if the sidebar is
shown in general outside the searching context of it.

Thus, it should be update if and only if the sidebar's
state changes due input unrelated to search. Or in other
words: If the state changes and it was not showing the
find sidebar, the setting should be update, otherwise
not.

Fixes: 6f7827aed766f1335ad9e2b0da25d2dfed31407e ("shell: Port PpsDocumentView to Rust")

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>