shell: document-view: use show-sidebar action to control sidebar's state
The sidebar_button uses the action as its toggle state.
When we initialise the sidebar's visibility, it should
be done through the action so that the action's state
and the sidebar's state never deviate.

Normally, the state is kept in sync by the signal handler
of notify::show-sidebar, but differences in the default
values can result in this discrepancy.

Adw.OverlaySplitView:show-sidebar defaults to true, while
our show-sidebar action defaulted to false. So,
technically the updated default value is a fix unto
itself for this. However, using the state directly here
ensures that the solution is more obvious as to why it
works, and that future changes to default values and
similar easily overlooked side effects are more
thoroughly mitigated.

Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/317

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>