libview: view: use zoom state of document model instead of local copy
The only reason for having a local copy of the zoom
level, is for ensuring queue_resize is not spammed, and
subsequent redraws triggering page re-rendering.

This only queues the resizing, reallocation, and
redrawing. GTK will do so for the next frame, so this is
capped by the screen's refresh rate, amongst other.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>