build: let nightly Flatpak crash on criticals
Criticals are a sign we are doing something severely
wrong. Making the nightly slightly less stable, I would
say, is something we can trade for more bug reports for
criticals, we are not experiencing ourselves.

There have been discussions about enabling this by
default in GLib. However, this suggestion here is, unlike
that discussion, not about production version of Papers.

Link: https://gitlab.gnome.org/GNOME/glib/-/issues/2715

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>