build: build Poppler from master for nightly
While we need to be careful with bumping the minimum
required Poppler version, we should also make sure newer
versions of Poppler do regress in ways we use it, most
notably in Poppler's cairo renderer, for which we cannot
rely on Okular users, or other users of the Qt bindings
to do the reporting for us.

Hopefully, we will then catch bugs in Poppler before it
hits a release, or at least we catch bugs in newer
Poppler release and have the option to withold a version
from our Flathub releases, unlike the recent incident,
where the broken signature box slipped into Poppler a
few versions before reported to us.

Link: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/348

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>